% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIQ.R
\name{BMIQ}
\alias{BMIQ}
\title{Beta MIxture Quantile dilation (BMIQ)}
\usage{
BMIQ(beta.v, design.v, nfit = 50000, th1.v = c(0.2, 0.75), niter = 5,
  tol = 0.001)
}
\arguments{
\item{beta.v}{A list of DNA methylation level}

\item{design.v}{A list of probe design('I' or 'II')}
}
\value{
A data frame contains normalized m and um, p value, and DNA methylation level
}
\description{
The BMIQ was developed by Andrew Teschendorff in 2013 and this implementation is the based on the version modified by Steve Horvath. I also removed and modified some parameters to simplify the function.
}
