% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIQ.R
\name{.BMIQ}
\alias{.BMIQ}
\title{Beta MIxture Quantile dilation (BMIQ)}
\usage{
.BMIQ(beta.v, design.v, nfit = 50000, th1.v = c(0.2, 0.75),
  niter = 5, tol = 0.001)
}
\arguments{
\item{beta.v}{A list of DNA methylation level}

\item{design.v}{A list of probe design('I' or 'II')}

\item{nfit}{number of probes of a given design type to use for the fitting. Default is 50000. Smaller values (~10000) will make BMIQ run faster at the expense of a small loss in accuracy. For most applications, 5000 or 10000 is ok.}

\item{th1.v}{thresholds used for the initialisation of the EM-algorithm, they should represent buest guesses for calling probes hemi-methylated and methylated, and will be refined by the EM algorithm. Default values work well in most cases.}

\item{niter}{maximum number of EM iterations to do. This number should be large enough to yield good fits to the type1 distribution. By default 5.}

\item{tol}{tolerance convergence threshold for EM algorithm. By default 0.001.}
}
\value{
A data frame contains normalized m and um, p value, and DNA methylation level
}
\description{
The BMIQ was developed by Andrew Teschendorff in 2013 and this implementation is the based on the version modified by Steve Horvath. I also removed and modified some parameters to simplify the function.
}
