% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmqn_BMIQ.R
\name{gmqn_bmiq}
\alias{gmqn_bmiq}
\title{GMQN + BMIQ}
\usage{
gmqn_bmiq(m, um, probe, type = "450k", ref = "default",
  verbose = TRUE)
}
\arguments{
\item{m}{The vector of methylation signal intensities.}

\item{um}{The vector of umethylation signal intensities.}

\item{probe}{the probe id for m and um.}

\item{type}{The type of methylation array which can be '450k'(default) or '850k'}

\item{ref}{The reference used to normalize data. By default, it uses the reference set up from GSE105018 which is also used in EWAS Datahub.}
}
\value{
A data frame contains normalized m and um, p value, and DNA methylation level
}
\description{
Using GMQN to conduct between array normalization and using BMIQ to conduct within array normalization. The BMIQ was developed by Andrew Teschendorff in 2013 and this implementation is the based on the version modified by Steve Horvath. I also removed and modified some parameters to simplify the function.
}
