% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmqn_normalize.R
\name{gmqn_normalize}
\alias{gmqn_normalize}
\title{Gaussian Mixture Quantile Normalization (GMQN)}
\usage{
gmqn_normalize(m, um, probe, type = "450k", ref = "default")
}
\arguments{
\item{m}{The vector of methylation signal intensities.}

\item{um}{The vector of umethylation signal intensities.}

\item{probe}{the probe id for m and um.}

\item{type}{The type of methylation array which can be '450k'(default) or '850k'}

\item{ref}{The reference used to normalize data. By default, it uses the reference set up from GSE105018 which is also used in EWAS Datahub.}
}
\value{
A data frame contains normalized m and um, p value, and DNA methylation level
}
\description{
The implementation of Gaussian Mixture Quantile Normalization (GMQN)
}
