% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmqn_normalize.R
\name{qmqn_normalize}
\alias{qmqn_normalize}
\title{Add together two numbers.}
\usage{
qmqn_normalize(m, um, probe, type = "450k",
  ref = list(t1.green.ref.mean = c(361.0165, 9646.7525), t1.green.ref.sd
  = c(278.9551, 5109.423), t1.red.ref.mean = c(696.0015, 12036.3621),
  t1.red.ref.sd = c(519.3141, 7002.863)))
}
\arguments{
\item{m}{The matrix of methylation signal intensities.}

\item{um}{The matrix of umethylation signal intensities.}

\item{probe}{the probe id for m and um.}

\item{type}{The type of methylation array which can be '450k'(default) or '850k'}

\item{ref}{The reference used to normalize data}
}
\value{
A data frame contains normalized m and um, p value, and DNA methylation level
}
\description{
Add together two numbers.
}
