% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_reference.R
\name{set_reference}
\alias{set_reference}
\title{Set reference for GMQN using user defined data}
\usage{
set_reference(m, um, type = "450k")
}
\arguments{
\item{m}{The data frame of methylation signal intensities.}

\item{um}{The data frame of umethylation signal intensities.}

\item{type}{The type of methylation array which can be '450k'(default) or '850k'}
}
\value{
A list contains referenced parameters
}
\description{
Set reference for GMQN using user defined data
}
